refactor details context
