+++ context/app/static/js/components/detailPage/BulkDataTransfer/BulkDataTransfer.jsx
@@ -1,8 +1,9 @@
 import React, { useState, useMemo, useCallback } from 'react';
 
 import SectionHeader from 'js/shared-styles/sections/SectionHeader';
-import { FilesContext, useDetailContext } from 'js/components/Contexts';
 import { DetailPageSection } from 'js/components/detailPage/style';
+import { useDetailContext } from 'js/components/detailPage/context';
+import FilesContext from 'js/components/detailPage/files/Files/context';
 import FileBrowserDUA from './FileBrowserDUA';
 import BulkDataTransferPanels from './BulkDataTransferPanels';
 import { StyledContainer } from './style';
