+++ context/app/static/js/components/detailPage/context.js
@@ -1,6 +1,10 @@
-import React from 'react';
+import React, { useContext } from 'react';
 
 // Stores the endpoints, `uuid`, and `hubmap_id` values.
 const DetailContext = React.createContext({});
 
-export default DetailContext;
+DetailContext.displayName = 'DetailContext';
+
+export { DetailContext };
+
+export const useDetailContext = () => useContext(DetailContext);
