+++ context/app/static/js/components/detailPage/files/Files/context.js
@@ -1,6 +1,10 @@
-import React from 'react';
+import React, { useContext } from 'react';
 
 // Stores the endpoints, `uuid`, and `hubmap_id` values.
 const FilesContext = React.createContext({});
 
-export default FilesContext;
+FilesContext.displayName = 'FilesContext';
+
+export { FilesContext };
+
+export const useFilesContext = () => useContext(FilesContext);
